requirements.txt

# THIS FILE IS READ-ONLY
# DO NOT MODIFY DEPENDENCIES
requests
python-whois
dnspython
phonenumbers
folium
pycountry
